chore: move streamlit frontend from streamlit to frontend:
