+++ src/rai_core/rai/frontend/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2024 Robotec.AI
+# Copyright (C) 2025 Robotec.AI
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .agents import wait_for_shutdown
-from .ros import ROS2Context
+from ..frontend.streamlit import run_streamlit_app
 
-__all__ = ["ROS2Context", "wait_for_shutdown"]
+__all__ = ["run_streamlit_app"]
